fix(util): rename utl directory
