--- src/util/console.js
@@ -1,3 +1,0 @@
-export function consoleWarn (message, vm) {
-  console.warn(`[Vuetify] ${message}`, vm)
-}
