+++ src/util/easing-patterns.js
