+++ src/util/mask.ts
