+++ src/util/rebuildFunctionalSlots.ts
